requirements.txt
anyio==4.12.1
asgiref==3.11.1
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
Django>=4.2,<5
easyocr==1.7.2
filelock==3.20.3
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.0
idna==3.11
ImageIO==2.28.1
Jinja2==3.1.6
joblib==1.5.3
lazy_loader==0.4
MarkupSafe==3.0.3
mpmath==1.3.0
networkx>=3.2,<3.5
ninja==1.13.0
nltk==3.9.2
numpy==1.26.4
opencv-python==4.7.0.72
opencv-python-headless==4.13.0.90
packaging==26.0
pillow==12.1.0
pyclipper==1.4.0
python-bidi==0.6.7
PyYAML==6.0.3
regex==2026.1.15
safetensors==0.7.0
scikit-image>=0.22,<0.26
scikit-learn==1.4.2
scipy==1.11.4
sentence-transformers==5.2.2
setuptools==80.10.2
shapely==2.1.2
shellingham==1.5.4
sqlparse==0.5.5
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2023.9.26
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.0.0
typer-slim==0.21.1
typing_extensions==4.15.0
tzdata==2025.3
